usabilidad-168
Mostrar información al usuario en index ahora funciona correctamente.
