+++ decide/base/urls.py
@@ -1,4 +1,3 @@
 from django.urls import path
-from base import views
 
 urlpatterns = []
