--- decide/base/views.py
@@ -1,6 +1,0 @@
-from django.shortcuts import render
-
-# Create your views here.
-
-def helpvoiceassistant(request):
-    return render(request,'helpvoiceassistant.html',)
