+++ decide/voting/urls.py
@@ -5,4 +5,5 @@ from . import views
 urlpatterns = [
     path('', views.VotingView.as_view(), name='voting'),
     path('<int:voting_id>/', views.VotingUpdate.as_view(), name='voting'),
+    path('getuser/', views.GetUserVotingsView.as_view())
 ]
